+++ source/utils/index.ts
@@ -1,0 +1,3 @@
+export { TsdError } from "./TsdError";
+export { isDiagnosticWithLocation } from "./isDiagnosticWithLocation";
+export { resolveCompilerOptions } from "./resolveCompilerOptions";