+++ source/utils/isDiagnosticWithLocation.ts
@@ -1,0 +1,5 @@
+import type * as ts from "@tsd/typescript";
+
+export const isDiagnosticWithLocation = (
+  diagnostic: ts.Diagnostic
+): diagnostic is ts.DiagnosticWithLocation => diagnostic.file !== undefined;