+++ tests/api.test.ts
@@ -1,3 +1,5 @@
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
+
 import { normalize, resolve } from "path";
 import { expect, test } from "@jest/globals";
 import tsd from "../";
@@ -7,12 +9,12 @@ test("returns `ts.SourceFile` object", () => {
   const { tsdResults } = tsd(fixturePath("failing"));
 
   expect(tsdResults).toHaveLength(2);
-  expect(normalize(tsdResults[0].file.fileName)).toEqual(
+  expect(normalize(tsdResults[0].file!.fileName)).toEqual(
     resolve("tests", "failing", "index.test.ts")
   );
 
-  expect(tsdResults[0].file.text).toEqual(tsdResults[1].file.text);
-  expect(tsdResults[0].file.text).toMatchInlineSnapshot(`
+  expect(tsdResults[0].file!.text).toEqual(tsdResults[1].file!.text);
+  expect(tsdResults[0].file!.text).toMatchInlineSnapshot(`
     "import { expectError, expectType } from \\"../../\\";
     import { makeDate } from \\".\\";